pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = [
    {name = "Mamchits Anastasia",email = "katara221995@mail.ru"}
]

readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.2"
pytest-cov = "^6.1.1"


[tool.poetry.group.lint.dependencies]
mypy = "^1.15.0"
black = "^25.1.0"
isort = "^6.0.1"


[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119
exclude = 'venv'